requirements.txt
backoff==2.2.1
certifi==2023.7.22
charset-normalizer==3.3.0
gql==3.4.1
graphql-core==3.2.3
idna==3.4
multidict==6.0.4
python-dotenv==1.0.0
requests==2.31.0
requests-toolbelt==1.0.0
urllib3==2.0.7
yarl==1.9.2
